requirements.txt
selenium==4.23.1
pytest==8.2.0
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
